repository-server/file_server/app/requirements.txt
streamlit==1.23.0
altair==4.2.0
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.31
pymysql==1.0.2
cryptography==39.0.1
aiofiles==23.1.0
python-multipart==0.0.6
mysql-connector-python==8.0.33
pandas==2.0.2
numpy==1.26.1
six==1.16.0
setuptools==58.0.0
httpx==0.24.0